+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### v1.2.9
 * Actor : `fire_` methods now default to `sNone` instead of `-1` if no sprite is provided.
     * Otherwise damage numbers do not show for clients.
+* Added Global class; this is just a shortform of `GM.variable_global_get/set`.
 
 ### v1.2.8
 * Callback, Callback_Raw : `add` can now accept numbers from `Callback.TYPE` enum.